Add javaEE annotation sample project (#1481)
* Delete AccountServlet.java

* Delete BankAppServletContextListener.java

* Delete LogInFilter.java

* Delete UploadCustomerDocumentsServlet.java

* Delete index.jsp

* Delete login.jsp

* Delete upload.jsp

* Delete web.xml

* Create javaeeannotations

* Delete javaeeannotations

* commit javaEE annotations project
