+++ jee7/src/main/java/com/baeldung/javaeeannotations/JavaEEAnnotationsSample/src/main/java/com/baeldung/javaeeannotations/AccountServlet.java
@@ -5,9 +5,6 @@ import java.io.PrintWriter;
 
 import javax.servlet.ServletConfig;
 import javax.servlet.ServletException;
-import javax.servlet.annotation.HttpConstraint;
-import javax.servlet.annotation.HttpMethodConstraint;
-import javax.servlet.annotation.ServletSecurity;
 import javax.servlet.annotation.WebInitParam;
 import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServletRequest;
@@ -19,20 +16,20 @@ import javax.servlet.http.HttpServletResponse;
   urlPatterns = {"/account", "/bankAccount" }, 
   initParams = { @WebInitParam(name = "type", value = "savings") }
   )
-@ServletSecurity(
+/*@ServletSecurity(
   value = @HttpConstraint(rolesAllowed = {"admin"}),
   httpMethodConstraints = {@HttpMethodConstraint(value = "POST", rolesAllowed = {"admin"})}
-  )
+  )*/
 public class AccountServlet extends javax.servlet.http.HttpServlet {
 
     String accountType = null;
 
-    @Override
     public void init(ServletConfig config) throws ServletException {
         accountType = config.getInitParameter("type");
     }
 
     public void doGet(HttpServletRequest request, HttpServletResponse response) throws IOException {
+
         PrintWriter writer = response.getWriter();
         writer.println("<html>Hello, I am an AccountServlet!</html>");
         writer.flush();
@@ -49,8 +46,8 @@ public class AccountServlet extends javax.servlet.http.HttpServlet {
 
         PrintWriter writer = response.getWriter();
         writer.println("<html> Balance of " + accountType + " account is: " + 
-           accountBalance + "<br> This account bares an interest rate of " + interestRate + 
-           " % </html>");
+            accountBalance + "<br> This account bares an interest rate of " + interestRate + 
+            " % </html>");
         writer.flush();
 
     }
