--- jee7/src/main/webapp/WEB-INF/web.xml
@@ -1,11 +1,0 @@
-<web-app xmlns="http://xmlns.jcp.org/xml/ns/javaee"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://xmlns.jcp.org/xml/ns/javaee
-                             http://xmlns.jcp.org/xml/ns/javaee/web-app_3_1.xsd" version="3.1">
-
-    <login-config>
-        <auth-method>BASIC</auth-method>
-        <realm-name>default</realm-name>
-    </login-config>
-
-</web-app>
