--- jee7/src/main/webapp/index.jsp
@@ -1,16 +1,0 @@
-<%@ page language="java" contentType="text/html; charset=ISO-8859-1"
-    pageEncoding="ISO-8859-1"%>
-<!DOCTYPE html PUBLIC "-//W3C//DTD HTML 4.01 Transitional//EN" "http://www.w3.org/TR/html4/loose.dtd">
-<html>
-<head>
-<meta http-equiv="Content-Type" content="text/html; charset=ISO-8859-1">
-<title>My Account</title>
-</head>
-<body>
-<form action="account" method="post">
-        Width: <input type="text" size="5" name="dep"/> 
-        &nbsp;&nbsp;
-        <input type="submit" value="Deposit" />
-</form>
-</body>
-</html>
